+++ src/syntax/ppx_sedlex.ml
@@ -234,10 +234,11 @@ let state_fun state = Printf.sprintf "__sedlex_state_%i" state
    returns the accepting rule index directly; otherwise it emits a function
    call to the generated state function. *)
 let call_state lexbuf (auto : Sedlex.dfa) state =
+  let loc = default_loc in
   let { Sedlex.trans; finals } = auto.(state) in
   if Array.length trans = 0 then (
     match best_final finals with
-      | Some i -> eint ~loc:default_loc i
+      | Some i -> [%expr Sedlexing.accept [%e lexbuf] [%e eint ~loc i]]
       | None -> assert false)
   else appfun (state_fun state) [lexbuf]