+++ test/codegen/test_gen.ml
@@ -27,11 +27,11 @@ let%expect_test "simple string match" =
       | _ -> Sedlexing.backtrack buf
     and __sedlex_state_1 buf =
       match __sedlex_partition_2 (Sedlexing.__private__next_int buf) with
-      | 0 -> 0
+      | 0 -> Sedlexing.accept buf 0
       | _ -> Sedlexing.backtrack buf
     and __sedlex_state_3 buf =
       match __sedlex_partition_3 (Sedlexing.__private__next_int buf) with
-      | 0 -> 0
+      | 0 -> Sedlexing.accept buf 0
       | _ -> Sedlexing.backtrack buf in
     match Sedlexing.start buf; __sedlex_state_0 buf with | 0 -> () | _ -> ()
     |}]
@@ -109,11 +109,11 @@ let%expect_test "multi-rule" =
        | _ -> Sedlexing.backtrack buf)
     and __sedlex_state_2 buf =
       match __sedlex_partition_3 (Sedlexing.__private__next_int buf) with
-      | 0 -> 0
+      | 0 -> Sedlexing.accept buf 0
       | _ -> Sedlexing.backtrack buf
     and __sedlex_state_4 buf =
       match __sedlex_partition_4 (Sedlexing.__private__next_int buf) with
-      | 0 -> 1
+      | 0 -> Sedlexing.accept buf 1
       | _ -> Sedlexing.backtrack buf in
     match Sedlexing.start buf; __sedlex_state_0 buf with
     | 0 -> ()
@@ -153,7 +153,7 @@ let%expect_test "as binding: simple" =
       | _ -> Sedlexing.backtrack buf
     and __sedlex_state_2 buf =
       match __sedlex_partition_3 (Sedlexing.__private__next_int buf) with
-      | 0 -> 0
+      | 0 -> Sedlexing.accept buf 0
       | _ -> Sedlexing.backtrack buf in
     match Sedlexing.start buf; __sedlex_state_0 buf with
     | 0 ->
@@ -235,7 +235,7 @@ let%expect_test "as binding: multiple bindings" =
       | _ -> Sedlexing.backtrack buf
     and __sedlex_state_2 buf =
       match __sedlex_partition_3 (Sedlexing.__private__next_int buf) with
-      | 0 -> 0
+      | 0 -> Sedlexing.accept buf 0
       | _ -> Sedlexing.backtrack buf in
     match Sedlexing.start buf; __sedlex_state_0 buf with
     | 0 ->
@@ -352,8 +352,8 @@ let%expect_test "as binding: shared prefix or-pattern" =
       | _ -> Sedlexing.backtrack buf
     and __sedlex_state_5 buf =
       match __sedlex_partition_6 (Sedlexing.__private__next_int buf) with
-      | 0 -> (Sedlexing.__private__set_mem_value buf 0 0; 0)
-      | 1 -> (Sedlexing.__private__set_mem_value buf 0 1; 0)
+      | 0 -> (Sedlexing.__private__set_mem_value buf 0 0; Sedlexing.accept buf 0)
+      | 1 -> (Sedlexing.__private__set_mem_value buf 0 1; Sedlexing.accept buf 0)
       | _ -> Sedlexing.backtrack buf in
     match Sedlexing.start buf;
           Sedlexing.__private__init_mem buf 1;
@@ -418,12 +418,13 @@ let%expect_test "as binding: 3-way or reuses disc cell" =
     and __sedlex_state_3 buf =
       match __sedlex_partition_4 (Sedlexing.__private__next_int buf) with
       | 0 -> (Sedlexing.__private__set_mem_value buf 0 0; __sedlex_state_4 buf)
-      | 1 -> (Sedlexing.__private__set_mem_value buf 0 1; 0)
+      | 1 -> (Sedlexing.__private__set_mem_value buf 0 1; Sedlexing.accept buf 0)
       | _ -> Sedlexing.backtrack buf
     and __sedlex_state_4 buf =
       Sedlexing.mark buf 0;
       (match __sedlex_partition_5 (Sedlexing.__private__next_int buf) with
-       | 0 -> (Sedlexing.__private__set_mem_value buf 0 2; 0)
+       | 0 ->
+           (Sedlexing.__private__set_mem_value buf 0 2; Sedlexing.accept buf 0)
        | _ -> Sedlexing.backtrack buf) in
     match Sedlexing.start buf;
           Sedlexing.__private__init_mem buf 1;
@@ -483,11 +484,11 @@ let%expect_test "as binding: multi-rule" =
       | _ -> Sedlexing.backtrack buf
     and __sedlex_state_1 buf =
       match __sedlex_partition_2 (Sedlexing.__private__next_int buf) with
-      | 0 -> 0
+      | 0 -> Sedlexing.accept buf 0
       | _ -> Sedlexing.backtrack buf
     and __sedlex_state_3 buf =
       match __sedlex_partition_3 (Sedlexing.__private__next_int buf) with
-      | 0 -> 1
+      | 0 -> Sedlexing.accept buf 1
       | _ -> Sedlexing.backtrack buf in
     match Sedlexing.start buf; __sedlex_state_0 buf with
     | 0 ->
@@ -525,7 +526,7 @@ let%expect_test "as binding: wrapping alternation" =
       | _ -> Sedlexing.backtrack buf
     and __sedlex_state_1 buf =
       match __sedlex_partition_2 (Sedlexing.__private__next_int buf) with
-      | 0 -> 0
+      | 0 -> Sedlexing.accept buf 0
       | _ -> Sedlexing.backtrack buf in
     match Sedlexing.start buf; __sedlex_state_0 buf with
     | 0 ->
@@ -913,8 +914,8 @@ let%expect_test "optim: dead tag elimination" =
     and __sedlex_state_2 buf =
       match __sedlex_partition_3 (Sedlexing.__private__next_int buf) with
       | 0 -> __sedlex_state_2 buf
-      | 1 -> 0
-      | 2 -> 1
+      | 1 -> Sedlexing.accept buf 0
+      | 2 -> Sedlexing.accept buf 1
       | _ -> Sedlexing.backtrack buf in
     match Sedlexing.start buf;
           Sedlexing.__private__init_mem buf 1;
@@ -1023,7 +1024,7 @@ let%expect_test "optim: tag remapping after coalescing" =
     and __sedlex_state_2 buf =
       match __sedlex_partition_3 (Sedlexing.__private__next_int buf) with
       | 0 -> __sedlex_state_2 buf
-      | 1 -> 0
+      | 1 -> Sedlexing.accept buf 0
       | _ -> Sedlexing.backtrack buf in
     match Sedlexing.start buf; __sedlex_state_0 buf with
     | 0 ->
@@ -1084,7 +1085,7 @@ let%expect_test "optim: set_prev with backtracking" =
     and __sedlex_state_2 buf =
       Sedlexing.mark buf 0;
       (match __sedlex_partition_2 (Sedlexing.__private__next_int buf) with
-       | 0 -> 0
+       | 0 -> Sedlexing.accept buf 0
        | 1 -> __sedlex_state_2 buf
        | _ -> Sedlexing.backtrack buf) in
     match Sedlexing.start buf;
@@ -1224,7 +1225,7 @@ let%expect_test "as binding: or-chain then nested or on right" =
       | _ -> Sedlexing.backtrack buf
     and __sedlex_state_3 buf =
       match __sedlex_partition_4 (Sedlexing.__private__next_int buf) with
-      | 0 -> (Sedlexing.__private__set_mem_value buf 0 0; 0)
+      | 0 -> (Sedlexing.__private__set_mem_value buf 0 0; Sedlexing.accept buf 0)
       | _ -> Sedlexing.backtrack buf
     and __sedlex_state_5 buf =
       match __sedlex_partition_5 (Sedlexing.__private__next_int buf) with
@@ -1245,7 +1246,7 @@ let%expect_test "as binding: or-chain then nested or on right" =
       | _ -> Sedlexing.backtrack buf
     and __sedlex_state_9 buf =
       match __sedlex_partition_7 (Sedlexing.__private__next_int buf) with
-      | 0 -> (Sedlexing.__private__set_mem_value buf 0 2; 0)
+      | 0 -> (Sedlexing.__private__set_mem_value buf 0 2; Sedlexing.accept buf 0)
       | _ -> Sedlexing.backtrack buf
     and __sedlex_state_11 buf =
       match __sedlex_partition_4 (Sedlexing.__private__next_int buf) with